streamline test resources
- central package for test data creation;
- made the testing code more DRY;
- reduce amount of traffic send to cockroachdb cluster,
because of high latency from my location;
- reduced required testing time >15 min to <3 minutes;
- all database int types are converted to bigint,
to provide portability between PostgreSQL and cockroachDB;

+++ internal/db/migrations/20221007154326_count_schema.up.sql
@@ -1,7 +1,7 @@
 create schema count;
 
 create table count.methods(
-  id serial primary key,
+  id bigserial primary key,
   method varchar not null,
   path varchar not null,
 
@@ -9,6 +9,6 @@ create table count.methods(
 );
 
 create table count.requests(
-  method_id int not null references count.methods(id),
+  method_id bigint not null references count.methods(id),
   request_timestamp timestamptz not null
 );
