+++ internal/db/migrations/20221015210435_create_daily_reports.up.sql
@@ -1,6 +1,6 @@
 create table count.daily_method_totals(
   day date not null default current_date,
-  method_id int not null references count.methods(id),
+  method_id bigint not null references count.methods(id),
   total bigint,
 
   primary key(day, method_id)
